# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API TP2 Partie 1",
        "version": "0.1"
    },
    "paths": {
        "/api/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "65b7ef1b1e076c033bf91055f53bd9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "email",
                                    "last_name",
                                    "first_name"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucessful"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user and return token",
                "operationId": "b5cee56757c3c79c5b5f115fdc33f460",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the current user",
                "operationId": "b46f8611fc7d1123f2a3d573cee661c4",
                "responses": {
                    "204": {
                        "description": "No content(successful)"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/critics": {
            "post": {
                "tags": [
                    "Critics"
                ],
                "summary": "Create a new critic",
                "operationId": "5152963e9c4be1ec4d595192a682ad53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score",
                                    "comment",
                                    "user_id",
                                    "film_id"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "film_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucessful"
                    },
                    "404": {
                        "description": "Movie or User Not Found"
                    },
                    "409": {
                        "description": "Existing Critic On Specific Movie"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/films": {
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Create a new film",
                "operationId": "ce1501bac824524d1e4706075650bf11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "release_year",
                                    "length",
                                    "description",
                                    "rating",
                                    "language_id",
                                    "special_features",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "release_year": {
                                        "type": "integer"
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    },
                                    "special_features": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "image": {
                                        "type": "string",
                                        "maxLength": 40
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Language Not Found"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/films/{film_id}": {
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Update an existing film",
                "operationId": "b650e9c8f4388d690532e75d2456f0d2",
                "parameters": [
                    {
                        "name": "film_id",
                        "in": "path",
                        "description": "film id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "release_year",
                                    "length",
                                    "description",
                                    "rating",
                                    "language_id",
                                    "special_features",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "release_year": {
                                        "type": "integer"
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    },
                                    "special_features": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "image": {
                                        "type": "string",
                                        "maxLength": 40
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Language Not Found"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/film/{film_id}": {
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Delete an existing film",
                "operationId": "aec0b07dd6898f6705d5d75eea2727d4",
                "parameters": [
                    {
                        "name": "film_id",
                        "in": "path",
                        "description": "film id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Film Not Found"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "operationId": "9cd1cee912e0176094a7ec83cc9b432f",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Critics",
            "description": "Critics"
        },
        {
            "name": "Films",
            "description": "Films"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}